requirements.txt
coverage==4.5.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.3
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
static3==0.7.0
Unipath==1.1
whitenoise==4.1.3
